test: Move Custom repositories tests to Repositories.test.tsx
This moves Custom repositories tests from `CreateImageWizard.content.test.tsx` to `Repositories.test.tsx`.
